backend/requirements.txt
[200~Flask==2.2.5
psycopg2-binary==2.9.9

